+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
+## 0.12.0
+
+BREAKING CHANGES
+
+- [core] Updated commission handling
+- [core] Fix multisend issue
+- [core] Extend max tx size
+- [api] Fixes in error responses
+
 ## 0.11.0
+
 BREAKING CHANGES
 
 - [core] Fix coin convert issue